test: DrawResultTest 생성
